Update index.adoc
I'm not sure what all changed from when the video was first created to now, but as I had spent 2 hours trying to figure out how to get this to run, I recalled running into this issue nearly a month ago when I first had the problem. Someone told me it was fine but also offered up this solution as well which DOES work on the https://stackblitz.com site.